+++ ft_memset.c
@@ -6,7 +6,7 @@
 /*   By: mstupnik <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/09/17 17:21:56 by mstupnik          #+#    #+#             */
-/*   Updated: 2019/09/19 13:39:57 by mstupnik         ###   ########.fr       */
+/*   Updated: 2019/09/19 14:37:13 by mstupnik         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	*ft_memset(void *b, int c, size_t len)
 {
-	int				x;
+	size_t			x;
 	unsigned char	*str;
 
 	x = 0;
